Refactor to use Token Struct

+++ contracts/interfaces/IHyperswapCallee.sol
@@ -3,5 +3,5 @@ pragma solidity ^0.8.13;
 
 
 interface IHyperswapCallee {
-    function hyperswapCall(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function hyperswapCall(address sender, uint256 amount0, uint256 amount1, bytes calldata data) external;
 }
